requirements.txt
# Anthropic
anthropic>=0.40.0

# Async HTTP
aiohttp>=3.9.0
httpx>=0.27.0

# Pydantic & settings
pydantic>=2.6.0
pydantic-settings>=2.2.0
python-dotenv>=1.0.0

# Scheduling
apscheduler>=3.10.4

# Data / storage
sqlalchemy>=2.0.0

# RAG - all free, all local
chromadb>=0.5.0
sentence-transformers>=3.0.0

# News feeds
feedparser>=6.0.11
gdeltdoc>=1.3.0

# Reddit (free API)
praw>=7.7.1

# Telegram alerts (free)
python-telegram-bot>=21.0

# Polymarket
py-clob-client>=0.15.0

# Web3 for Polygon wallet
web3>=6.15.0

# Dashboard
streamlit>=1.35.0
plotly>=5.22.0
pandas>=2.2.0

# Utilities
loguru>=0.7.2
tenacity>=8.3.0
rich>=13.7.0
